SourceCode/requirements.txt
python==3.10.14
torch==2.3.1
numpy
pillow
imageio
matplotlib
scikit-image
pandas
natsort
tqdm
openpyxl